+++ pkg/ruler/remote/transport.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package remotequerier
+package remote
 
 import (
 	"flag"
